pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "feature_store_utils"
version = "0.0.3"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "A utility to generate ML features from yaml"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'python-dotenv',
    'databricks-feature-store',
    'Jinja2',
    'pandas',
    'pyaml'
    ]

[tool.setuptools.package-data]
"features.templates" = ["*.j2"]

[tool.pytest_env]
DATABRICKS_CLUSTER_ID = {value = "1127-14505-wla3xgc5", skip_if_set = true}

